icu-monitoring-system/ai-module/python/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
kafka-python==2.0.2
python-dotenv==1.0.0
httpx==0.26.0
